멤버스 적용 방 옵션별로 나누기

+++ src/main/java/egovframework/example/sample/web/Pot.java
@@ -11,7 +11,7 @@ public class Pot {
 	
 	private ArrayList<JSONObject> PotUserSeats = new ArrayList<JSONObject>();	
 	
-	public int JoinPot(int seat, int amount)
+	public long JoinPot(int seat, long amount)
 	{
 		if( amount <= 0 )
 		{
@@ -90,7 +90,7 @@ public class Pot {
 		return totalPot;
 	}
 	
-	public Pot PotSlit(int amount)
+	public Pot PotSlit(long amount)
 	{
 		Pot slitpot = new Pot();
